backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.37
asyncpg==0.30.0
alembic==1.14.1
pydantic-settings==2.7.1
bcrypt==5.0.0
cryptography==44.0.0
redis==5.2.1
httpx==0.28.1
PyJWT[crypto]==2.10.1
